Internal types

Strategic period types (TwoLevelTree)

Single types

TimeStruct.AbstractTreeNode — Type

AbstractTreeNode{S,T} <: AbstractStrategicPeriod{S,T}

Abstract base type for all tree nodes within a [TwoLevelTree] type.

TimeStruct.StratNode — Type

struct StratNode{S, T, OP<:TimeStructure{T}} <: AbstractTreeNode{S,T}

A structure representing a single strategic node of a [TwolevelTree]. It is created through iterating through StratTreeNodes.

It is equivalent to a StrategicPeriod of a TwoLevel time structure when utilizing a [TwolevelTree].

TimeStruct.StrategicScenario — Type

struct StrategicScenario

Desription of an individual strategic scenario. It includes all strategic nodes corresponding to a scenario, including the probability. It can be utilized within a decomposition algorithm.

Iterator types

TimeStruct.AbstractTreeStructure — Type

AbstractTreeStructure

Abstract base type for all tree timestructures within a [TwoLevelTree] type.

TimeStruct.StratTreeNodes — Type

struct StratTreeNodes{T, OP} <: AbstractTreeStructure

Type for iterating through the individual strategic nodes of a TwoLevelTree. It is automatically created through the function strat_periods, and hence, strategic_periods.

Iterating through StratTreeNodes using the WithPrev iterator changes the behaviour, although the meaining remains unchanged.

Strategic period types (TwoLevel)

Single types

TimeStruct.AbstractStrategicPeriod — Type

AbstractStrategicPeriod{S,T} <: TimeStructure{T}

Abstract base type for all strategic periods returned when iterating through the periods using the strat_periods function.

TimeStruct.StrategicPeriod — Type

StrategicPeriod <: AbstractStrategicPeriod

Time period for iteration of strategic periods of a TwoLevel time structure.

Iterator types

Representative period types

Single types

TimeStruct.AbstractRepresentativePeriod — Type

AbstractRepresentativePeriod

Abstract type used for time structures that represent a representative period.

TimeStruct.RepresentativePeriod — Type

RepresentativePeriod

A structure representing a single representative period supporting iteration over its time periods.

TimeStruct.StratNodeReprPeriod — Type

struct StratNodeReprPeriod{T,OP<:TimeStructure{T}} <: AbstractRepresentativePeriod{T}

A structure representing a single representative period of a StratNode of a [TwolevelTree]. It is created through iterating through StratNodeReprPeriods.

It is equivalent to a [StratReprPeriod] of a TwoLevel time structure when utilizing a [TwolevelTree].

Iterator types

TimeStruct.StratNodeReprPeriods — Type

struct StratNodeReprPeriods <: AbstractTreeStructure

Type for iterating through the individual presentative periods of a StratNode. It is automatically created through the function repr_periods.

Operational scenarios types

Single types

TimeStruct.OperationalScenario — Type

struct OperationalScenario

A structure representing a single operational scenario supporting iteration over its time periods.

TimeStruct.StratOperationalScenario — Type

struct StratOperationalScenario

A structure representing a single operational scenario for a strategic period supporting iteration over its time periods.

TimeStruct.StratNodeOperationalScenario — Type

struct StratNodeOperationalScenario{T,OP<:TimeStructure{T}}  <: AbstractOperationalScenario{T}

A structure representing a single operational scenario for a strategic node supporting iteration over its time periods. It is created through iterating through StratNodeOpScens.

It is equivalent to a StratOperationalScenario of a TwoLevel time structure when utilizing a [TwolevelTree].

TimeStruct.StratNodeReprOpscenario — Type

struct StratNodeReprOpscenario{T} <: AbstractOperationalScenario{T}

A structure representing a single operational scenario for a representative period in A [TwoLevelTree] structure supporting iteration over its time periods.

Iterator types

TimeStruct.StrategicScenarios — Type

struct StrategicScenarios

Type for iteration through the individual strategic scenarios represented as [StrategicScenario].

TimeStruct.StratNodeOpScens — Type

struct StratNodeOpScens <: AbstractTreeStructure

Type for iterating through the individual operational scenarios of a StratNode. It is automatically created through the function opscenarios.

TimeStruct.StratNodeReprOpscenarios — Type

struct StratNodeReprOpscenarios <: AbstractTreeStructure

Type for iterating through the individual operational scenarios of a StratNodeReprPeriod. It is automatically created through the function opscenarios.

Operational period types

TimeStruct.TimePeriod — Type

abstract type TimePeriod

Abstract type used for a uniform interface for iterating through time structures and indexing of time profiles.

TimeStruct.SimplePeriod — Type

struct SimplePeriod <: TimePeriod

A single time period returned when iterating through a SimpleTimes structure

TimeStruct.CalendarPeriod — Type

struct CalendarPeriod <: TimePeriod

A single time period returned when iterating through a CalendarTimes structure with duration measured in hours (by default).

TimeStruct.OperationalPeriod — Type

struct OperationalPeriod <: TimePeriod

Time period for iteration of a TwoLevel time structure.

TimeStruct.TreePeriod — Type

struct TreePeriod{P} <: TimePeriod where {P<:TimePeriod}

Time period for iteration of a TwoLevelTree time structure. This period has in addition to an operational period also the two fields branch and prob_branch corresponding to the respective branch and probability of the branch

Using OperationalScenarios

The probability will always only correspond to the branch probability, even when you utilize OperationalScenarios. Using the function probability includes however the scenario probability.