Internal types

Strategic period types (TwoLevelTree)

Single types

TimeStruct.AbstractTreeNode — Type

AbstractTreeNode{S,T} <: AbstractStrategicPeriod{S,T}

Abstract base type for all tree nodes within a TwoLevelTree type.

TimeStruct.StratNode — Type

struct StratNode{S, T, OP<:TimeStructure{T}} <: AbstractTreeNode{S,T}

A structure representing a single strategic node of a TwoLevelTree. It is created through iterating through StratTreeNodes.

It is equivalent to a StrategicPeriod of a TwoLevel time structure when utilizing a TwoLevelTree.

TimeStruct.StrategicScenario — Type

struct StrategicScenario

Desription of an individual strategic scenario. It includes all strategic nodes corresponding to a scenario, including the probability. It can be utilized within a decomposition algorithm.

Iterator types

TimeStruct.AbstractTreeStructure — Type

AbstractTreeStructure

Abstract base type for all tree timestructures within a TwoLevelTree type.

TimeStruct.StratTreeNodes — Type

struct StratTreeNodes{S, T, OP} <: AbstractTreeStructure

Type for iterating through the individual strategic nodes of a TwoLevelTree. It is automatically created through the function strat_periods, and hence, strategic_periods.

Iterating through StratTreeNodes using the WithPrev iterator changes the behaviour, although the meaning remains unchanged.

TimeStruct.StrategicScenarios — Type

struct StrategicScenarios

Type for iteration through the individual strategic scenarios represented as StrategicScenario.

Strategic period types (TwoLevel)

Single types

TimeStruct.AbstractStrategicPeriod — Type

AbstractStrategicPeriod{S,T} <: TimeStructure{T}

Abstract type used for time structures that represent a strategic period. These periods are obtained when iterating through the strategic periods of a time structure declared by the function strat_periods.

TimeStruct.SingleStrategicPeriod — Type

SingleStrategicPeriodWrapper{T,SP<:TimeStructure{T}} <: AbstractStrategicPeriod{T,T}

A type representing a single strategic period supporting iteration over its time periods. It is created when iterating through SingleStrategicPeriodWrapper.

TimeStruct.StrategicPeriod — Type

StrategicPeriod{S,T,OP<:TimeStructure{T}} <: AbstractStrategicPeriod{S,T}

A type representing a single strategic period supporting iteration over its time periods. It is created when iterating through StratPeriods.

Iterator types

TimeStruct.SingleStrategicPeriodWrapper — Type

SingleStrategicPeriodWrapper{T,SP<:TimeStructure{T}} <: TimeStructure{T}

Type for iterating through the individual strategic periods of a time structure without TwoLevel. It is automatically created through the function strat_periods.

TimeStruct.StratPeriods — Type

StratPeriods{S,T,OP}

Type for iterating through the individual strategic periods of a TwoLevel time structure. It is automatically created through the function strat_periods.

Representative period types

Single types

TimeStruct.AbstractRepresentativePeriod — Type

AbstractRepresentativePeriod{T} <: TimeStructure{T}

Abstract type used for time structures that represent a representative period. These periods are obtained when iterating through the representative periods of a time structure declared by the function repr_periods.

TimeStruct.SingleReprPeriod — Type

SingleReprPeriod{T,OP<:TimeStructure{T}} <: AbstractRepresentativePeriod{T}

A type representing a single representative period supporting iteration over its time periods. It is created when iterating through SingleReprPeriodWrapper.

TimeStruct.RepresentativePeriod — Type

RepresentativePeriod{T,OP<:TimeStructure{T}} <: AbstractRepresentativePeriod{T}

A type representing a single representative period supporting iteration over its time periods. It is created when iterating through ReprPeriods.

TimeStruct.StratReprPeriod — Type

StratReprPeriod{T,OP<:TimeStructure{T}} <: AbstractRepresentativePeriod{T}

A type representing a single representative period supporting iteration over its time periods. It is created when iterating through StratReprPeriods.

TimeStruct.StratNodeReprPeriod — Type

struct StratNodeReprPeriod{T,OP<:TimeStructure{T}} <: AbstractRepresentativePeriod{T}

A structure representing a single representative period of a StratNode of a TwoLevelTree. It is created through iterating through StratNodeReprPeriods.

It is equivalent to a StratReprPeriod of a TwoLevel time structure when utilizing a TwoLevelTree.

Iterator types

TimeStruct.SingleReprPeriodWrapper — Type

SingleReprPeriodWrapper{T,OP<:TimeStructure{T}} <: TimeStructure{T}

Type for iterating through the individual representative periods of a time structure without RepresentativePeriods. It is automatically created through the function repr_periods.

TimeStruct.ReprPeriods — Type

ReprPeriods{S,T,OP}

Type for iterating through the individual representative periods of a RepresentativePeriods time structure. It is automatically created through the function repr_periods.

TimeStruct.StratReprPeriods — Type

StratReprPeriods{OP}

Type for iterating through the individual representative periods of a StrategicPeriod time structure. It is automatically created through the function repr_periods.

TimeStruct.StratNodeReprPeriods — Type

struct StratNodeReprPeriods <: AbstractTreeStructure

Type for iterating through the individual presentative periods of a StratNode. It is automatically created through the function repr_periods.

Operational scenarios types

Single types

TimeStruct.AbstractOperationalScenario — Type

AbstractOperationalScenario{T} <: TimeStructure{T}

Abstract type used for time structures that represent an operational scenario. These periods are obtained when iterating through the operational scenarios of a time structure declared by the function opscenarios.

TimeStruct.SingleScenario — Type

SingleScenario{T,SC<:TimeStructure{T}} <: AbstractRepresentativePeriod{T}

A type representing a single operational scenario supporting iteration over its time periods. It is created when iterating through SingleScenarioWrapper.

TimeStruct.OperationalScenario — Type

OperationalScenario{T,OP<:TimeStructure{T}} <: AbstractOperationalScenario{T}

A type representing a single operational scenario supporting iteration over its time periods. It is created when iterating through OpScens.

TimeStruct.ReprOperationalScenario — Type

ReprOperationalScenario{T,OP<:TimeStructure{T}} <: AbstractOperationalScenario{T}

A type representing a single operational scenarios supporting iteration over its time periods. It is created when iterating through RepOpScens.

TimeStruct.StratOperationalScenario — Type

StratOperationalScenario{T,OP<:TimeStructure{T}} <: AbstractOperationalScenario{T}

A type representing a single operational scenario supporting iteration over its time periods. It is created when iterating through StratOpScens.

TimeStruct.StratReprOpscenario — Type

StratReprOpscenario{T, OP<:TimeStructure{T}} <: AbstractRepresentativePeriod{T}

A type representing a single representative period supporting iteration over its time periods. It is created when iterating through StratReprPeriods.

TimeStruct.StratNodeOperationalScenario — Type

struct StratNodeOperationalScenario{T,OP<:TimeStructure{T}}  <: AbstractOperationalScenario{T}

A structure representing a single operational scenario for a strategic node supporting iteration over its time periods. It is created through iterating through StratNodeOpScens.

It is equivalent to a StratOperationalScenario of a TwoLevel time structure when utilizing a TwoLevelTree.

TimeStruct.StratNodeReprOpScenario — Type

struct StratNodeReprOpScenario{T} <: AbstractOperationalScenario{T}

A structure representing a single operational scenario for a representative period in A TwoLevelTree structure supporting iteration over its time periods.

Iterator types

TimeStruct.SingleScenarioWrapper — Type

SingleScenarioWrapper{T,OP<:TimeStructure{T}} <: TimeStructure{T}

Type for iterating through the individual operational scenarios of a time structure without OperationalScenarios. It is automatically created through the function opscenarios.

TimeStruct.OpScens — Type

OpScens{T,OP}

Type for iterating through the individual operational scenarios of a OperationalScenarios time structure. It is automatically created through the function opscenarios.

TimeStruct.RepOpScens — Type

RepOpScens{OP}

Type for iterating through the individual operational scenarios of a RepresentativePeriod time structure. It is automatically created through the function opscenarios.

TimeStruct.StratOpScens — Type

StratOpScens{OP}

Type for iterating through the individual operational scenarios of a StrategicPeriod time structure. It is automatically created through the function opscenarios.

TimeStruct.StratReprOpscenarios — Type

StratReprOpscenarios{OP}

Type for iterating through the individual operational scenarios of a StrategicPeriod time structure with RepresentativePeriods. It is automatically created through the function opscenarios.

TimeStruct.StratNodeOpScens — Type

struct StratNodeOpScens <: AbstractTreeStructure

Type for iterating through the individual operational scenarios of a StratNode. It is automatically created through the function opscenarios.

TimeStruct.StratNodeReprOpScens — Type

struct StratNodeReprOpScens <: AbstractTreeStructure

Type for iterating through the individual operational scenarios of a StratNodeReprPeriod. It is automatically created through the function opscenarios.

Operational period types

TimeStruct.TimePeriod — Type

abstract type TimePeriod

Abstract type used for a uniform interface for iterating through time structures and indexing of time profiles.

TimeStruct.SimplePeriod — Type

struct SimplePeriod{T<:Number} <: TimePeriod

Time period for a single operational period. It is created through iterating through a SimpleTimes time structure.

TimeStruct.CalendarPeriod — Type

struct CalendarPeriod <: TimePeriod

Time period for a single operational period. It is created through iterating through a CalendarTimes time structure with duration measured in hours (by default).

TimeStruct.ScenarioPeriod — Type

ScenarioPeriod{P} <: TimePeriod where {P<:TimePeriod}

Time period for a single operational period. It is created through iterating through a OperationalScenarios time structure. It is as well created as period within OperationalPeriod when the time structure includes OperationalScenarios.

TimeStruct.ReprPeriod — Type

ReprPeriod{P} <: TimePeriod where {P<:TimePeriod}

Time period for a single operational period. It is created through iterating through a RepresentativePeriods time structure. It is as well created as period within OperationalPeriod when the time structure includes RepresentativePeriods.

TimeStruct.OperationalPeriod — Type

struct OperationalPeriod <: TimePeriod

Time period for a single operational period. It is created through iterating through a TwoLevel time structure.

TimeStruct.TreePeriod — Type

struct TreePeriod{P} <: TimePeriod where {P<:TimePeriod}

Time period for iteration of a TwoLevelTree time structure. This period has in addition to an operational period also the two fields branch and prob_branch corresponding to the respective branch and probability of the branch

Using OperationalScenarios

The probability will always only correspond to the branch probability, even when you utilize OperationalScenarios. Using the function probability includes however the scenario probability.